Highlight active modifiers
In EventHandler::loop(), we iterate through the modifiers, and if any of
them is active, we set them to the appropriate color. Otherwise, we turn
their color off.

We do this in loop(), so we can easily catch one-shot modifiers too, or
modifiers pressed through a macro, and so on, without having to peek
into the various events we delegate to other components.

Fixes #4.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>