Use a base color map to colorize the layers
Rework how colors are set: we have a map of the current base colors,
just indexes, not full cRGB colors. This is the base, that is applied
every cycle. We blend other colors onto this on various events. The
blended colors are not recorded anywhere, they'll turn back to the base
color upon release anyway.

This fixes #9, hopefully.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>