Turn F11 into a tap-dance key
Since R0C0 (which F11 is assigned to) is also the programmer key, I can't just
hold it when I want to flash, because then I'd send a held F11 to the host,
which I do not want to. Yet, I still want F11 in that place.

So I made it into a tap-dance key, which acts as F11 on first tap, and as a
dummy afterwards. This is perfect for my use cases.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>