Move the numbers off the base layer
Inspired by an idea from @nfnty, the numbers were moved from the top row on the
base layer to a numpad-like layout on a separate layer. The symbols previously
available when shifting these keys are now there without the shifting.

As such, the `ShapeShifter` plugin is no longer in use.

Fixes #24.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>