Small key rearrangement
The main reason for the move is that I need access to the `GUI` key from the
`NAV` layer. To this end, `GUI` has been moved to the left side, beside `Z`. The
`play/pause` key that was there, was moved to the right side, beside `J`. And
the `LEAD` key was therefore moved to the thumb cluster, where `GUI` used to be.

A similar change was applied to the `ADORE` layer too.

Images were updated, and the colormap now highlights `LEAD` and `GUI` too.

Fixes #40.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>