Rearrange a few keys on the ADORE layer
Because my pinkies do not like reaching the row above the home row, move X from
there to the bottom row, swapping it with Quote.

Furthermore, because I want to move my index fingers less, move B to the left
side, to where Comma used to be, which shifted one key to the right, knocking
Period over to the right side, where B used to be.

The idea is that Comma and Period being on different hands will lead to better
alteration, and B being below the home position will mean less index finger
movement to reach it.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>