Drop the EmptyLayerForceOff hack
This was required in the past, to work around a bug of some sort, I guess. It is
certainly not needed now, and it caused the `_EMPTY` layer to be forced off all
the time.

With this gone, and Kaleidoscope not requiring layer keys to be transparent
anymore, we can simplify our custom `getKey` implementation too.

Fixes #5.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>