Migrate the PROGMEM->EEPROM keymap transfer to host-side
Instead of having a special Focus command to transfer a layer from PROGMEM to
KEYMAP, move that code to the host side. This saves plenty of space for us, and
removes a lot of unnecessary code from the firmware. Also shows how to do some
funky stuff with the the Focus protocol.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>